Refactored dpos components and updated imports
