+++ src/modules/dpos/validator/components/registerDelegateMultiStep/element.js
