+++ src/modules/dpos/validator/components/registerDelegateMultiStep/index.js
