+++ src/modules/dpos/validator/components/registerDelegateMultiStep/navigator.js
