+++ src/modules/dpos/validator/components/registerDelegateMultiStep/navigatorButton.js
