+++ src/modules/dpos/validator/components/registerDelegateMultiStep/utils.js
