+++ src/modules/dpos/validator/components/warnPunishedDelegate/index.js
