+++ src/modules/dpos/validator/components/warnPunishedDelegate/voteWarning.js
