+++ src/modules/dpos/validator/components/warnPunishedDelegate/warnPunishedDelegate.css
